split_path_and_profile() move to CalculationTarget::parse()
