+++ src/lib.rs
@@ -10,25 +10,12 @@ use std::{
 use self::filemode::FileMode;
 use anyhow::{anyhow, Context};
 use git2::{Oid, Repository};
-use regex::Regex;
 use sver_config::CalculationTarget;
 
 pub struct Version {
     pub repository_root: String,
     pub path: String,
     pub version: String,
-}
-
-fn split_path_and_profile(value: &str) -> CalculationTarget {
-    let regex = Regex::new("(.+):([a-zA-Z0-9-_]+)").unwrap();
-    let caps = regex.captures(value);
-    caps.map(|caps| {
-        CalculationTarget::new(
-            caps.get(1).unwrap().as_str().to_string(),
-            caps.get(2).unwrap().as_str().to_string(),
-        )
-    })
-    .unwrap_or_else(|| CalculationTarget::new(value.to_string(), "default".to_string()))
 }
 
 fn relative_path(repo: &Repository, path: &Path) -> anyhow::Result<PathBuf> {
@@ -87,29 +74,4 @@ fn find_repository(from_path: &Path) -> anyhow::Result<Repository> {
         }
     }
     Err(anyhow!("repository was not found"))
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{split_path_and_profile, sver_config::CalculationTarget};
-
-    #[test]
-    fn test_split() {
-        assert_eq!(
-            split_path_and_profile("hello"),
-            CalculationTarget::new("hello".to_string(), "default".to_string())
-        );
-        assert_eq!(
-            split_path_and_profile("hello:world"),
-            CalculationTarget::new("hello".to_string(), "world".to_string())
-        );
-        assert_eq!(
-            split_path_and_profile(r"c:\hello"),
-            CalculationTarget::new(r"c:\hello".to_string(), "default".to_string())
-        );
-        assert_eq!(
-            split_path_and_profile(r"c:\hello:world-wide"),
-            CalculationTarget::new(r"c:\hello".to_string(), "world-wide".to_string())
-        );
-    }
 }
